Add the initial logic to support displaying a "brush"
Added a second canvas that is positioned above the shared canvas, this canvas can be
used to display things like the current active 'brush' (currently just a circle).

The radius of the circle being displayed is tied to the number the user selects for
the size of the line they are drawing

The color of the line on the outside edge of the circle is set to '#777' to try and
offset it from most colors that would be chosen, some logic may want to be added to
prevent that line from matching the color "under" it .. that may be more difficult than
it is worth. The circle is filled, the color of that fill changes based on the current
color selected by the user.